Create GalleriesLayout, move main from pages to it
