+++ src/app/galleries/architecture/ArchitectureVertImages.js
