+++ src/app/galleries/layout.js
@@ -1,0 +1,5 @@
+import styles from "./layout.module.css";
+
+export default function GalleriesLayout({ children }) {
+  return <main className={styles.main}>{children}</main>;
+}
