+++ src/app/galleries/layout.module.css
@@ -1,0 +1,18 @@
+.main {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: stretch;
+  height: calc(var(--vh, 1vh) * 100);
+  width: calc(var(--vw, 1vw) * 100);
+  position: relative;
+  background-color: #272727;
+}
+
+/* Mobile landscape gallery */
+
+@media only screen and (max-width: 949px) and (orientation: landscape) {
+  .main {
+    display: block;
+  }
+}
