+++ src/app/galleries/travel/page.module.css
@@ -1,14 +1,3 @@
-.main {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: stretch;
-  height: calc(var(--vh, 1vh) * 100);
-  width: calc(var(--vw, 1vw) * 100);
-  position: relative;
-  background-color: #272727;
-}
-
 .rotate-message-container {
   height: fit-content;
   width: 100vw;
@@ -102,9 +91,6 @@
 /* Mobile landscape gallery */
 
 @media only screen and (max-width: 949px) and (orientation: landscape) {
-  .main {
-    display: block;
-  }
 
   .gallery {
     height: 100vh;
